--- modules/control/conf/chassis.conf
@@ -1,3 +1,0 @@
---v=4
---arduino_device_name=ttyArduino
---speed_feedback=1
