+++ modules/control/conf/control.conf
@@ -1,5 +1,7 @@
 --flagfile=/apollo/modules/common/data/global_flagfile.txt
 --control_conf_file=/apollo/modules/control/conf/control_conf.pb.txt
+--magnetic_enable=false
+--desired_v=1
 --enable_speed_station_preview=false
 --enable_interpolation_by_time=false
 --use_preview_speed_for_table=false
